Deduplication of type casting codes
