+++ c10/util/Half.h
@@ -389,34 +389,6 @@ struct scalar_value_type<ComplexHalf> {
   using type = Half;
 };
 
-// The old implementation of Converter as a function made nvcc's head explode
-// when we added std::complex on top of the specializations for CUDA-only types
-// like __half, so I rewrote it as a templated class (so, no more overloads,
-// just (partial) specialization).
-
-template <typename To, typename From, typename Enable = void>
-struct Converter {
-  To operator()(From f) {
-    return static_cast<To>(f);
-  }
-};
-
-template <typename To, typename From>
-To convert(From from) {
-  return Converter<To, From>()(from);
-}
-
-template <typename To, typename FromV>
-struct Converter<
-    To,
-    std::complex<FromV>,
-    typename std::enable_if<
-        guts::negation<is_complex_t<To>>::value>::type> {
-  To operator()(std::complex<FromV> f) {
-    return static_cast<To>(f.real());
-  }
-};
-
 // In some versions of MSVC, there will be a compiler error when building.
 // C4146: unary minus operator applied to unsigned type, result still unsigned
 // C4804: unsafe use of type 'bool' in operation
@@ -501,18 +473,6 @@ typename std::enable_if<is_complex_t<From>::value, bool>::type overflows(
       overflows<
              typename scalar_value_type<To>::type,
              typename From::value_type>(f.imag());
-}
-
-template <typename To, typename From>
-To checked_convert(From f, const char* name) {
-  // Converting to bool can't overflow so we exclude this case from checking.
-  if (!std::is_same<To, bool>::value && overflows<To, From>(f)) {
-    std::ostringstream oss;
-    oss << "value cannot be converted to type " << name
-        << " without overflow: " << f;
-    throw std::domain_error(oss.str());
-  }
-  return convert<To, From>(f);
 }
 
 C10_API std::ostream& operator<<(std::ostream& out, const Half& value);
